Restored NYPLConfiguration
This class currently provides the main feed URL for the catalog. Said URL had
been moved to "Localizable.strings", but that wasn't correct as the language
for the catalog should not be tied to the language used for the rest of the
application. The approach for getting the main feed URL will change in the
future: This is simply a temporary measure until requirements are known.
be tied to the language used for the rest of the app. As such,