Fixed minor issue with progress reporting when redownloading a book
Previously, starting a download would not issue a notification from the
download center. This was fine if the book had not been downloaded during the
current application run, but if it had been, views would begin to show the book
as downloading with the progress still at 100% from the previous download.