Issue #121 Implementation of chapter/book reading progress indicators using a progress bar and a description label the estimated positions are calculated by combining pagination information coming from readium and manual byte calculations of the actual xml page sizes. chapter progress information is based on the pagination information received from Readium and may not be 100% accurate on the last "page" of a chapter overal book reading process is based on total bytes length of all xml pages of the book and chapter progress information and may not be 100% accurate on the last "page" of a book
the information only shows up when user activates the "overhead ui display" and is updated on each page turn whether the ui is displayed or not. it is also recalculated and updated after TOC navigation events.
chapter numbering is based on the spine item index and may not match the spine item titles seen in the TOC

the reading progress view updates it's colour scheme based on the overall colour scheme setting (white | black | sepia)