Implemented new sign out procedure
It is now possible for the user to sign out even if they are not able to
deauthorize their device, a case that usually would happen if their PIN had
changed and thus their credentials were no longer valid. An internet
connectivity check was inserted to ensure that deauthorization will not be
skipped simply because the user's device is not online.