Make .xcconfig files understandable to CocoaPods
Prior to this commit, we had four `*.xcconfig` files:

* `Simplified.debug.xcconfig`, which just included `Simplified.xcconfig`
and `Simplified+RMSDK.xcconfig`;

* `Simplified.release.xcconfig`, which was identical to
`Simplified.debug.xcconfig`;

* `Simplified.xcconfig`, which contained the bulk of the project
configuration;

* `Simplified+RMSDK.xcconfig`, which included headers for DRM-related
libraries and set `FEATURE_DRM_CONNECTOR=1`.

The project was set to use the `Simplified.debug.xcconfig` and
`Simplified.release.xcconfig` configurations for debug and release
respectfully. The SimplyE target _within_ debug and release was set to
use `Simplified+RMSDK.xcconfig`, however.

It was the later portion that caused CocoaPods to issue the following
warnings:

    [!] CocoaPods did not set the base configuration of your project
    because your project already has a custom config set. In order for
    CocoaPods integration to work at all, please either set the base
    configurations of the target `SimplyE` to `Pods/Target Support
    Files/Pods-SimplyE/Pods-SimplyE.debug.xcconfig` or include the
    `Pods/Target Support Files/Pods-SimplyE/Pods-SimplyE.debug.xcconfig`
    in your build configuration (`Simplified+RMSDK.xcconfig`).

    [!] CocoaPods did not set the base configuration of your project
    because your project already has a custom config set. In order for
    CocoaPods integration to work at all, please either set the base
    configurations of the target `SimplyE` to `Pods/Target Support
    Files/Pods-SimplyE/Pods-SimplyE.release.xcconfig` or include the
    `Pods/Target Support
    Files/Pods-SimplyE/Pods-SimplyE.release.xcconfig` in your build
    configuration (`Simplified+RMSDK.xcconfig`).

The following fixed all of the above:

* I moved all configuration into `Simplified.xcconfig`.

* I set the project to use `Simplified.xcconfig` for both debug and
release.

* I removed the superfluous `Simplified.debug.xcconfig` and
`Simplified.release.xcconfig`.

* In the project settings, I removed the configuration setting for the
target _within_ debug and release. This allowed CocoaPods to set it
itself which resolved the warnings above.

This commit does not resolve long-standing issues in which DRM-free
builds are not particularly easy, but it should not make them any worse
either. I updated "README.md" to reflect the change: You must now
comment out the last line of "Simplified.xcconfig" instead of deleting
"Simplified+RMSDK.xcconfig".