SIMPLY-2072 Clean up Background Fetch tasks
Remove scheduling of background task from Background Fetch call, since by definition we are already in the background (even if the method runs on the main thread). Also, the whole point of Background Fetch is to run the callback code as fast as possible and in less than 30s.

Verified that the BackgroundFetch completionHandler is called in less than 4 seconds (iPhone 11 on wifi). Before removing backgroud task scheduling it was called 1 second later.

TODO: understand why we need to start a background task during background fetch, since we are already in the background.