SIMPLY-2675 Fix retain cycle during NYPLRemoteViewController init
The crash's stack trace points at an issue inside a NYPLSession callback during a NSLog call, but nothing there seems like an actual culprit. However NYPLSession may be called from NYPLOpenSearchDescription, which may be called by NYPLCatalogSearchViewController. The latter is the last screen view of most of the crashes. Since the crash is a EXC_BAD_ACCESS, this points to a issue with invalid memory.

So this may not fix the crash, but it is fixing a retain cycle which may lead to problems down the line.