Create new target + workspace to debug R2-related issues
The new SimplyE-R2dev target duplicates the main SimplyE target and changes how R2Shared, R2Streamer, R2Navigator are integrated. Those frameworks are now embedded from the built products of the workspace. So one has to build them first manually instead of relying on Carthage.

As a consequence those dependencies will have to be re-added to the Cartfile once the feature/readium2 branch is merged down.

Note: in the new target R2Streamer is embedded via the carthage copy-framework phase to mimic what the r2-testapp is doing to try to see if that made a difference in how the OpenDyslexic font is handled.