SIMPLY-2724 Refactor card creation objects using factory methods
This fixes a regression introduced on the feature branch where the regular configuration was retained for the lifetime of the businessLogic object, while on `develop` it was created fresh each time. This is even more important for the juvenile coordinator which may retain state related to the user (e.g. the parent barcode)..