Use authorization_identifier for juvenile APIs
The barcode field in NYPLUserAccount may contain either a barcode or the username, but the Platform APIs for juvenile accounts require to specify what you're passing in (either a barcode or username), but we on the client don't know what the user has signed in with.

For NYPL accounts, and only for NYPL Accounts, the NYPLUserAccount::authorization_identifier instead always contains a barcode: this can be a 14-digit NYPL-issued barcode, or a 16-digit NYC ID (issued by the city of NY). This was confirmed by @leonardr.