SIMPLY-2843 When not loading, make sure to show Try Again UI
I inspected the codebase to verify why we RemoteVC is receiving nil URLs occasionally, and it seems like this happens at navigation controller initialization time. So in that case, showing a reload/error UI would be incorrect, because it's not really an error. These situations are resolved by creating a new VC with the URL received from server (e.g.) once the authentication document loads. The previously created VC is replaced and dealloc'ed.