SIMPLY-2843 Log every sign in/out error + provide better user messages
I did a review of both Sign In VCs and logged every single error that can originate from both.
Also made sure to log the (hashed) barcode since it won't be otherwise available in Crashlytics if sign-in fails.
This also adds an alert before performing sign-out to warn the user that book downloads/returns authorizations or book registry syncing is in progress.