SIMPLY-2843 Reload libraries if catalog URL is nil, then reload Catalog
The Catalog screen is dependent on the presence of a catalog URL that comes from the  library authentication document. So if that's missing, we need to first reload the library accounts from the AccountsManager otherwise we'd be stuck in the water.

This is a big encapsulation breakage but it's necessary in the current architecture.