SIMPLY-2881 Add parsing of Problem Documents in network layer
The Problem Document info is returned up in additional userInfo keys in the NSError passed in the completion closure.

Currently the new network layer is only used for loading the authentication document and the list of libraries. While in both cases the upper level code does not make usage of the returned errors, it will now certainly receive Errors in case a Problem Document is receieved.