SIMPLY-2975 Surface rationale of download errors in Crashlytics reports
Expose a more detailed reason of error in the summary of the Crashlytics report (which coincides with the `NSError.domain`) and use `NSError.errorCode` as a way to group related errors together. E.g. all the business-logic-related book download errors now have the same error code, but the `context` aka NSError.domain will expose the distributor, the book medium  and reason of the error. This will create a larger number of non-fatal "issues" in Crashlytics but each one will contain less instances. This also addresses the fact that it's not possible to link to one event instance inside the same "issue"/group.