SIMPLY-3104 Correct launch sequence after migrating
The key part here is setting the currentAccountId for OE early on. This is ok since for OE we only have that one library. In SimplyE I think this was being set up as part of the default Simplified library.

Also add a lot of debug checkpoints to help diagnose launch sequence issues.