SIMPLY-2510 Refactor sign-in logic handling DRM authorization
Note that the 2 implementations of `drmAuthorizeWithUsername:password:loggingContext:` where slightly different in the 2 VCs, with the modal implementation performing an additional check to warn the user if the DRM authorization takes too long. Since this seems like a good idea, this is now added to both (the VC dismissal will result in nothing for a non-modal VC in the stack).