SIMPLY-2510 Refactor business logic delegate methods
Rewrite the delegate methods from the point of view of the business logic. Move some logging logic that was in the deleegate implementation to the business logic. Rewrite the URLSession challenge handler in swift to facilitate integration.

This way it was possible to clear out some more state from the VCs and simplify the api with less back'n'forth between business logic and VC.