SIMPLY-3390 Prevent crash for Unknown Right Mgmt
If server returns Content-Length == 0 for a fulfillment request, the download task does not download anything and goes straight to the `didFinishDownloadingToURL:` callback. However in this case the book's `rightsManagement` has not been set yet, which defaults to the "unknown" case.

Also make the "alert user" logic a bit more readable, making explicit when we have to present an alert or not.