SIMPLY-2401 Consolidate auth challenge handlers
This also fixes a probable bug where we were rejecting the auth challenge a priori if the uiDelegate has no username/passwd (in NYPLSignInURLSessionChallengeHandler), even for auth methods such as OAuth/SAML that do not work with username/passwd.

Another probable bug was the fact that it was possible to `reject` the `NSURLAuthenticationMethodServerTrust` authentication method, which is something related to certificate validation (e.g. in SSL/TLS handshake): `performDefaultHAndling` seems more appropriate than `reject`. During testing, from the user POV no difference was noted whether we used one or the other.