SIMPLY-2401 Refactor sign-in/out to use centralized networking
This allows to process problem documents more consistently.

Because authentication requests prefer no caching, this also refactors the Executor/Responder to allow customization of the caching behavior.
Auth logic also requires a way to provide credentials to the responder to answer auth challenges while the user is not fully signed in: see NYPLBasicAuthCredentialsProvider.
The auth challenge logic now is refactored to an object to facilitate reuse in conjunction with different credentials providers.