SIMPLY-2401 Fix inconsistency in order of execution
Other parts of the login flow check to see if we are on the main thread already before async'ing, and if we are not consistent in that we may end up in a place where we think everything has been updated, but something is still in flux. This was highlighted by a unit test added in previous commit.