Change branch reference for r2-navigator
In order to get the fixes for race conditions when restoring the reading progress in our fork of r2-navigator-swift, we will need to bring its `develop` branch up-to-date with upstream. However, that will break the current SimplyE build. To address this, I pushed a new `develop-before-CP` branch on our fork that matches the same commit we currently use (51a8dc1154fad6ad206bda00c8f3c4002cbada0d). This way we can refer to that in the Cartfile (instead of `develop`) on the current SimplyE builds on `develop`.

Note: this is an interim solution because at the moment there's no official release of R2 that includes our fix for race conditions when restoring the reading progress.