IOS-338 Prevent fastlane to use 2FA on CI jobs
Fastlane docs list `SPACESHIP_ONLY_ALLOW_INTERACTIVE_2FA` as an env var to set in a CI context to prevent it to use 2FA for authenticating with Apple. This is confirmed by the code, which instead will raise an exception:
https://github.com/fastlane/fastlane/blob/e874a47c6e2e0e61590a03d3b71e75e5a505d1ce/spaceship/lib/spaceship/two_step_or_factor_client.rb#L7

Note that `FASTLANE_IS_INTERACTIVE` is not a variable we can set: fastlane sets that automatically.