Prevent possible crash from implicitly unwrapped optional
This is a hack temporary fix. Granted we should NOT create a dummy cell, but since this would otherwise be preventing access to a critical section of the app, and we currently do not have time to address this, for now this is better than crashing.

I only experienced this situation if we were using the QA servers and then switched to not using them.