OE-514 Fix race condition between updating keychain and BookRegistry sync
Saving on the keychain happens asynchronously and may take a while. In the mean time we do not want the sign-in logic to think that we are 100% authenticated and proceed to update the BookRegistry. By exposing the updateQueue where the keychain logic works, we can have the sign-in logic use the same queue and properly sequence both things.