OE-514 Fix authDefinition getter to return the actual state of the property
Without this we would get a nonnull value when actually we have no authDefinition, for example when we are signed out. This would confuse the sign-in logic to think that we might be already signed in; for example, `hasOAuthClientCredentials()` could incorrectly return `true` even when we are actually signed out.