OE-514 Restore logic to return a default authDefinition when signed-out
In the previous checkin of NYPLUserAccount, the authDefinition getter was modified to always and only return the value saved in the keychain. However this broke the need to obtain the authentication definition when signed out. This checkin restores the pre-existing logic in a separate/explicit getter, even though that had some fragility issues.