OE-586 Detect selectedAuthentication based on credentials
There's a situation where we are signed in but somehow the business logic has no selected authentication. This has been proved difficult to reproduce reliably, but it seems to happen occasionally when upgrading from 2.4.0 to 2.5.0. This is probably related to the change in the getter for `NYPLUserAccount::authDefinition` introduced in b66b227 on 6.22/2022, which was however necessary, especially for OE who has multiple ways of signing in.